fix nt found in movie delete

+++ src/app/movies/movies.controller.ts
@@ -12,6 +12,7 @@ import {
   UseInterceptors,
   ClassSerializerInterceptor,
   CacheInterceptor,
+  NotFoundException,
 } from '@nestjs/common';
 import { MoviesService } from './movies.service';
 import { CreateMovieDto } from './dto/create-movie.dto';
@@ -56,6 +57,8 @@ export class MoviesController {
   @Delete(':id')
   @HttpCode(204)
   async remove(@Param('id', ParseIntPipe) id: number): Promise<void> {
+    const movie = this.findOne(id);
+    if (!movie) throw new NotFoundException('Movie Not found');
     this.moviesService.remove(id);
     return;
   }
